refactor: shop&recent 위젯 분리중
