+++ src/pages/noticedetail/RecentPostsWidget/RecentPostsWidget.styled.ts
@@ -1,0 +1,33 @@
+import styled from "styled-components";
+
+export const RecentWrap = styled.div`
+  margin: 60px 0 120px 0;
+`;
+
+export const BigText = styled.h1`
+  color: var(--The-julge-black, #111322);
+  font-size: 28px;
+  font-weight: 700;
+  letter-spacing: 0.56px;
+  margin-bottom: 24px;
+`;
+
+export const PostContainer = styled.div`
+  // NoticePostList.styled로 분리
+  /* display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  column-gap: 14px;
+  row-gap: 32px;
+  margin-top: 20px;
+
+  @media (max-width: 1200px) {
+    grid-template-columns: repeat(2, 1fr);
+  }
+
+  @media (max-width: 768px) {
+    grid-template-columns: 1fr 1fr;
+  } */
+`;
+
+
+
