--- src/pages/noticedetail/data/mockNoticeData.ts
@@ -1,23 +1,0 @@
-export const mockNoticeData = {
-  "item": {
-    "id": "2ad3ac93-0054-442e-a882-d6cce8c10470",
-    "hourlyPay": 1234456,
-    "startsAt": "2023-07-19",
-    "workhour": 12,
-    "description": "텍스트가 길어지면 더보기 뜨게 수정하기 텍스트가 길어지면 더보기 뜨게 수정하기 텍스트가 길어지면 더보기 뜨게 수정하기 텍스트가 길어지면 더보기 뜨게 수정하기 텍스트가 길어지면 더보기 뜨게 수정하기",
-    "closed": false,
-    "shop": {
-      "item": {
-        "id": "63fcc375-5d0a-4ba4-ac5b-101b03973c74",
-        "name": "에바네 한식❤️",
-        "category": "한식",
-        "address1": "서울시 중구",
-        "address2": "왜..",
-        "description": "테스트",
-        "imageUrl": "https://bootcamp-project-api.s3.ap-northeast-2.amazonaws.com/0-1/the-julge/1970eebc-dfaf-4fea-8b86-071c28c3b78d-%EC%9D%B4%EB%AA%A8%EC%A7%80.png",
-        "originalHourlyPay": 10000
-      },
-      "href": "/api/0-1/the-julge/shops/63fcc375-5d0a-4ba4-ac5b-101b03973c74"
-    }
-  },
-}
