+++ src/pages/noticedetail/index.tsx
@@ -1,117 +1,23 @@
-import { useEffect, useState } from 'react';
-import Image from 'next/image';
-import { useRouter } from 'next/router';
+import { useState } from 'react';
 import * as S from './index.styled';
-import { useSearchParams } from 'next/navigation';
 import Header from '@/shared/components/Header/Header';
 import Footer from '@/shared/components/Footer/Footer';
-import PostInform from '@/shared/components/Post/PostInform/PostInform';
-import CustomModal from './component/CustomModal/CustomModal';
-import CustomButton from '@/shared/components/Button/CustomButton/CustomButton';
-import RedButton from '@/shared/components/Button/RedButton/RedButton';
-import cautionImg from '@/assets/caution.svg';
-import checkImg from '@/assets/check.svg';
+import ShopDetailWidget from './ShopDetailWidget/ShopDetailWidget';
+import DetailWidget from './DetailWidget/DetailWidget';
+import RecentPostsWidget from './RecentPostsWidget/RecentPostWidget';
 import { useModal } from '@/shared/store/useModal';
-import { useUserQuery } from '@/models/user/useUserData';
-import PostPrice from '@/shared/components/Post/PostPrice/PostPrice';
-import { formatWorkTime } from './utils/useTimeUtils';
-import { useHandleModal } from './utils/useHandleModal';
-import { useGetNoticeDetail } from '../../models/notice/useGetNoticeDetail';
-import { renderSpinner } from '@/shared/utils/renderSpinner';
-import PostList from '@/shared/components/Post/PostList/PostList';
+import { useHandleModal, HandleModalType } from './utils/useHandleModal';
 import { NoticeData } from '@/shared/types/post';
 
 const NoticeDetail = () => {
-  const searchParams = useSearchParams();
-  const shop_id = searchParams.get('shop_id');
-  const notice_id = searchParams.get('notice_id');
-  const {
-    data: userData,
-    isError: userError,
-    isLoading: userLoading,
-  } = useUserQuery();
-  const {
-    data: noticeData,
-    isError: noticeError,
-    isLoading: noticeLoading,
-  } = useGetNoticeDetail(shop_id || '', notice_id || '');
-
-  const { setIsOpen, key, isOpen } = useModal();
+  const { key, isOpen } = useModal();
   const [isApplied, setIsApplied] = useState(false);
-  const router = useRouter();
-  const [recentPosts, setRecentPosts] = useState<NoticeData[]>([]);
 
-  useEffect(() => {
-    if (typeof window !== 'undefined') {
-      const recentPostsFromStorage = localStorage.getItem('recentPosts');
-      recentPostsFromStorage
-        ? setRecentPosts(JSON.parse(recentPostsFromStorage))
-        : setRecentPosts([]);
-    } else {
-      setRecentPosts([]);
-    }
-  }, []);
+  // useHandleModal 훅을 사용하여 handleModal 객체를 가져옴
+  const { applyClick, cancelClick, confirmCancel, closeCancelModal, confirm } = useHandleModal({ setIsApplied });
 
-  useEffect(() => {
-    if (notice_id && noticeData && typeof window !== 'undefined') {
-      const saveRecentPost = (postId: string) => {
-        setRecentPosts((prevPosts: NoticeData[]) => {
-          const shopData = noticeData.data.item.shop.item;
-
-          const newPost: NoticeData = {
-            item: {
-              id: notice_id,
-              hourlyPay: noticeData.data.item.hourlyPay,
-              startsAt: noticeData.data.item.startsAt,
-              workhour: noticeData.data.item.workhour,
-              description: noticeData.data.item.description,
-              closed: noticeData.data.item.closed,
-              shop: {
-                item: {
-                  id: shopData.id,
-                  name: shopData.name,
-                  category: shopData.category,
-                  address1: shopData.address1,
-                  address2: shopData.address2,
-                  description: shopData.description,
-                  imageUrl: shopData.imageUrl,
-                  originalHourlyPay: shopData.originalHourlyPay,
-                },
-                href: noticeData.data.item.shop.href,
-              },
-            },
-            links: [],
-          };
-
-          const updatedPosts = [
-            newPost,
-            ...(prevPosts || []).filter(
-              (post: NoticeData) => post.item.id !== notice_id,
-            ),
-          ].slice(0, 6);
-
-          localStorage.setItem('recentPosts', JSON.stringify(updatedPosts));
-          return updatedPosts;
-        });
-      };
-
-      saveRecentPost(notice_id);
-    }
-  }, [notice_id, noticeData]);
-
-  const handleModal = useHandleModal({ userData, setIsApplied });
-
-  const modalHeader =
-    key === 'profileAlert' ? (
-      <>
-        <Image src={cautionImg} alt="경고 표시" />내 프로필을 먼저 등록해주세요.
-      </>
-    ) : key === 'applySuccess' ? (
-      <>
-        <Image src={checkImg} alt="체크 표시" />
-        신청이 완료되었습니다.
-      </>
-    ) : null;
+  // NoticeData는 실제 데이터로 대체되어야 함
+  const noticeData = {} as NoticeData;
 
   if (!noticeData) {
     return <div>Loading...</div>;
@@ -122,129 +28,19 @@ const NoticeDetail = () => {
       <Header />
       <S.PageLayout>
         <S.SmallText>식당</S.SmallText>
-        <S.TextWrap>
-          <S.BigText>{noticeData.data.item.shop.item.name}</S.BigText>
-        </S.TextWrap>
-
-        <S.ContextWrap>
-          <S.ImageContainer>
-            <img
-              src={noticeData.data.item.shop.item.imageUrl}
-              alt={noticeData.data.item.shop.item.name}
-            />
-          </S.ImageContainer>
-
-          <S.TextContainer>
-            <S.SmallText>시급</S.SmallText>
-            <S.PriceWrap>
-              <PostPrice
-                isClosed={false}
-                defaultHourlyPay={noticeData.data.item.hourlyPay}
-                currentHourlyPay={
-                  noticeData.data.item.shop.item.originalHourlyPay
-                }
-              />
-            </S.PriceWrap>
-            <S.WidgetWrap>
-              <PostInform
-                isClosed={false}
-                type="시간"
-                content={formatWorkTime(
-                  noticeData.data.item.startsAt,
-                  noticeData.data.item.workhour,
-                )}
-              />
-              <PostInform
-                isClosed={false}
-                type="장소"
-                content={noticeData.data.item.shop.item.address1}
-              />
-            </S.WidgetWrap>
-            <S.DetailText>
-              <p>{noticeData.data.item.shop.item.description}</p>
-            </S.DetailText>
-            {isApplied ? (
-              <div style={{ width: '346px' }}>
-                <CustomButton
-                  onClick={handleModal.cancelClick}
-                  color="#EA3C12"
-                  text="취소하기"
-                />
-              </div>
-            ) : (
-              <S.CustomRedButton onClick={handleModal.applyClick}>
-                신청하기
-              </S.CustomRedButton>
-            )}
-          </S.TextContainer>
-        </S.ContextWrap>
-
-        <S.DescripContainer>
-          <S.SmallText isBlack>공고 설명</S.SmallText>
-          <p>{/* 공백 */}</p>
-          <p>{noticeData.data.item.description}</p>
-        </S.DescripContainer>
-
-        <S.RecentWrap>
-          <S.BigText>최근에 본 공고</S.BigText>
-          <S.PostContainer>
-            {renderSpinner(<PostList items={recentPosts} />, noticeLoading)}
-          </S.PostContainer>
-        </S.RecentWrap>
+        <ShopDetailWidget
+          noticeData={noticeData}
+          isApplied={isApplied}
+          handleModal={{ applyClick, cancelClick, confirmCancel, closeCancelModal, confirm } as HandleModalType} // 타입 단언을 사용하여 HandleModalType으로 형 변환
+        />
+        <DetailWidget noticeData={noticeData} />
+        <RecentPostsWidget
+          isOpen={isOpen}
+          key={key}
+          handleModal={{ applyClick, cancelClick, confirmCancel, closeCancelModal, confirm } as HandleModalType} // 타입 단언을 사용하여 HandleModalType으로 형 변환
+        />
       </S.PageLayout>
       <Footer />
-      <CustomModal
-        modalKey={key}
-        modalHeader={modalHeader}
-        modalFooter={
-          key === 'profileAlert' ? (
-            <div style={{ width: '80px' }}>
-              <CustomButton
-                text="확인"
-                color="#EA3C12"
-                onClick={handleModal.confirm}
-              />
-            </div>
-          ) : (
-            <div style={{ width: '80px' }}>
-              <CustomButton
-                text="확인"
-                color="#EA3C12"
-                onClick={handleModal.confirm}
-              />
-            </div>
-          )
-        }
-      />
-
-      {isOpen && key === 'cancelModal' && (
-        <CustomModal
-          modalKey="cancelModal"
-          modalHeader={
-            <>
-              <Image src={checkImg} alt="체크 표시" />
-              <p>신청을 취소하시겠어요?</p>
-            </>
-          }
-          modalFooter={
-            <>
-              <div style={{ width: '80px' }}>
-                <CustomButton
-                  onClick={handleModal.closeCancelModal}
-                  text="아니오"
-                  color="#EA3C12"
-                />
-              </div>
-              <div style={{ width: '80px' }}>
-                <RedButton
-                  onClick={handleModal.confirmCancel}
-                  text="취소하기"
-                />
-              </div>
-            </>
-          }
-        />
-      )}
     </>
   );
 };
