+++ src/pages/noticedetail/utils/useHandleModal.ts
@@ -1,22 +1,32 @@
+// useHandleModal.ts
 import { useRouter } from 'next/router';
 import { useModal } from '@/shared/store/useModal';
+import { useUserData } from '@/shared/store/useUserData';
 import { useState } from 'react';
-import { UseQueryResult } from 'react-query';
+
+// 타입 정의
+export type HandleModalType = {
+  applyClick: () => void;
+  cancelClick: () => void;
+  confirmCancel: () => void;
+  closeCancelModal: () => void;
+  confirm: () => void;
+};
 
 interface HandleModalProps {
-  userData: UseQueryResult<any, unknown>;
   setIsApplied: (value: boolean) => void;
 }
 
 export const useHandleModal = ({
-  userData,
   setIsApplied,
-}: HandleModalProps) => {
+}: HandleModalProps): HandleModalType => {
   const router = useRouter();
   const { setIsOpen, setIsClose, key } = useModal();
+  const userData = useUserData(); // 이곳에서 유저 데이터를 사용한다고 가정
 
   const applyClick = () => {
-    if (!userData?.data.item.phone) {
+    if (!userData?.phone) {
+      // 예시로 userData의 phone 필드가 없으면 profileAlert를 엽니다
       setIsOpen('profileAlert');
     } else {
       setIsOpen('applySuccess');
