+++ src/shared/store/useUserData.ts
@@ -2,6 +2,7 @@ import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
 
 interface Props {
+  phone: any;
   user_id: string;
   type: 'employee' | 'employer' | '';
   setUserId: (newId: string) => void;
